Added Coroutines for enemy and player

+++ Assets/Scripts/Scriptable Objects/BulletScriptable Object/BulletScriptableObject.cs
@@ -3,11 +3,12 @@ using UnityEngine;
 
 [CreateAssetMenu (fileName = "BulletScriptableObject", menuName = "ScriptableObject/NewBulletScriptableObject")]
 public class BulletScriptableObject : ScriptableObject 
-{
-   public BulletType bulletType;
+{  
+   public BulletType bulletType ;
    public GameObject bulletPrefab;
    public float bulletSpeed; 
-   public float damage;
+   public int damage;
+
 }
